requirements.txt
flask==2.3.3
pandas==2.0.3
groq==0.4.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
python-dotenv==1.0.0
streamlit==1.34.0
gunicorn==21.2.0
